Move transcrypt python files to the template dir
